+++ src/main/java/edu/csupomona/cs480/models/SongNote.java
@@ -1,7 +1,7 @@
 package edu.csupomona.cs480.models;
 
 public class SongNote {
-	public int songNoteId;
+	public long songNoteId;
 	public int sequenceNum;
 	public int startBeat;
 	public int duration;
